+++ cjs/core.js
@@ -1,0 +1,12 @@
+"use strict";
+exports.__esModule = true;
+exports.__includes = exports.LABELS = exports.opts = void 0;
+exports.opts = {
+    enumerable: false,
+    writable: true
+};
+exports.LABELS = [];
+function __includes(arr, item) {
+    return arr.indexOf(item) !== -1;
+}
+exports.__includes = __includes;