+++ index.js
@@ -1,5 +1,0 @@
-import def, { createTimer as ct } from './cjs/index.js';
-export { default as promises } from './cjs/promises.js';
-
-export const Timer = def;
-export const createTimer = ct;