+++ index.mjs
@@ -1,0 +1,4 @@
+import cjsModule from './cjs/index.js';
+/*eslint no-empty-pattern: "warn"*/
+export const { promises, createTimer } = cjsModule;
+export default cjsModule;