+++ tests/main.test.js
@@ -1,11 +1,11 @@
-const Timer = require('../cjs/index.js').Timer;
+const Timer = require('../cjs/timer');
 
 describe('Timer class test suite', () => {
 	test('it must measure no time between calls', () => {
 		for (let i = 0; i < 20; i++) {
 			const timer = new Timer().start().stop();
 			const time = timer.getDifference();
-			expect(time).toBe(0);
+			expect(time).toBeLessThanOrEqual(1);
 		}
 	});
 	test('getDifference and computeDifference should produce identical output', () => {
@@ -20,21 +20,21 @@ describe('Timer class test suite', () => {
 	describe('error handling', () => {
 		const timer = new Timer();
 		test('throws an error if you attempt to start twice', () => {
-			expect(() => timer.start().start()).toThrowError(Error);
+			expect(() => timer.start().start()).toThrow(Error);
 		});
 		test('throws an error if you attempt to stop twice', () => {
-			expect(() => timer.stop().stop()).toThrowError(Error);
+			expect(() => timer.stop().stop()).toThrow(Error);
 		});
 		test('throws an error if you try to operate after closure', () => {
-			expect(() => timer.close().start()).toThrowError(Error);
-			expect(() => timer.close().stop()).toThrowError(Error);
+			expect(() => timer.close().start()).toThrow(Error);
+			expect(() => timer.close().stop()).toThrow(Error);
 		});
 		test('throws an error if specified a non-string as a label', () => {
-			expect(() => new Timer(5)).toThrowError(TypeError);
-			expect(() => new Timer(true)).toThrowError(TypeError);
-			expect(() => new Timer(false)).toThrowError(TypeError);
-			expect(() => new Timer({})).toThrowError(TypeError);
-			expect(() => new Timer([])).toThrowError(TypeError);
+			expect(() => new Timer(5)).toThrow(TypeError);
+			expect(() => new Timer(true)).toThrow(TypeError);
+			expect(() => new Timer(false)).toThrow(TypeError);
+			expect(() => new Timer({})).toThrow(TypeError);
+			expect(() => new Timer([])).toThrow(TypeError);
 		});
 		test('throws an error if tried to make more than one timer with the same label', () => {
 			expect(() => {